client/pyproject.toml
[tool.poetry]
name = "dj-grpc-benchmark"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
grpcio-tools = "^1.43.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
